Update ShaderPrecompiler.gd to work with 4.3
As it was, it triggers the error:

Invalid call. Nonexistent function 'get_surface_material_count' in base 'MeshInstance3D'.

If I´m understanding the code, the "for" in line 36 looks for the materials of the mesh, but since the "mesh" is a MeshInstance3D(line 34) and not a proper mesh the code checks for any surface_override_material and then checks for the proper mesh materials.

Was this intended? or is just a change in Godot 4.3